+++ apps/billets-web/app/blog/[series]/[slug]/page.tsx
@@ -1,34 +1,37 @@
-'use client';
-import { Suspense, use } from 'react';
+import { Suspense, cache, use } from 'react';
 
 import { GlobalErrorBoundaryRegistry } from '@/libs/registries/global-error-boundary-registry';
+import type { FetchGetSeriesItemSearchParams } from 'app/api/blog/series/[slug]/types';
+import { fetchGetSeriesItem } from 'app/blog/(fetchers)';
 import { LogDetailRenderer } from 'app/blog/(notion-render)/log-detail-renderer';
-import { AppLocaleSchema } from 'app/blog/(types)/i18n';
 import type { SeriesCategory } from 'app/blog/(types)/series';
-import { parseAsStringLiteral, useQueryState } from 'nuqs';
 
-const appLocaleParser = parseAsStringLiteral(
-  AppLocaleSchema.options // -> readonly ['ko','en']
-).withDefault('ko');
+const getSeriesItemStatic = cache(
+  async (slug: string, searchParams: FetchGetSeriesItemSearchParams) => {
+    const response = await fetchGetSeriesItem(slug, searchParams);
+    return response;
+  }
+);
 
-export default function SeriesSlugPage(props: {
+export default async function SeriesSlugPage(props: {
   params: Promise<{
     slug: string;
     series: SeriesCategory;
   }>;
 }) {
-  const params = use(props.params);
-  const [lang, setLang] = useQueryState('lang', appLocaleParser);
+  const params = await props.params;
+  const initialData = await getSeriesItemStatic(params.slug, {
+    seriesCategory: params.series,
+    appLocale: 'ko',
+  });
 
   return (
     <GlobalErrorBoundaryRegistry>
       <Suspense>
         <LogDetailRenderer
           slug={params.slug}
-          locale={lang}
           seriesCategory={params.series}
-          onClickAppLocale={setLang}
-          onClickBack={() => setLang((prev) => (prev === 'ko' ? 'en' : 'ko'))}
+          initialData={initialData}
         />
       </Suspense>
     </GlobalErrorBoundaryRegistry>